requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
pypdf
pyPDF2
faiss-cpu
python-dotenv
tiktoken
fastapi
aiofiles
uvicorn
jinja2
python-multipart

-e .